🚚 Rename: 컴포넌트 폴더구조 변경
