+++ src/components/Post/PostComment.jsx
@@ -3,9 +3,9 @@ import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import Btn from '../../assets/icons/icon-more.svg';
 import defaultImage from '../../assets/images/basic-profile-s.png';
-import BottomSheetModal from '../Modal/BottomSheetModal';
-import BottomSheetContent from '../Modal/BottomSheetContent';
-import ConfirmModal from '../Modal/ConfirmModal';
+import BottomSheetModal from '../../layouts/Modal/BottomSheetModal';
+import BottomSheetContent from '../../layouts/Modal/BottomSheetContent';
+import ConfirmModal from '../../layouts/Modal/ConfirmModal';
 import authInstance from '../../api/instance/authInstance';
 import useModal from '../../hooks/useModal';
 import { convertTime } from '../../utils/convertTime';
