+++ src/components/Profile/ProfilePosts.jsx
@@ -2,12 +2,12 @@ import styled from 'styled-components';
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import authInstance from '../../api/instance/authInstance';
-import PostCard from '../Post/PostCard';
+import PostCard from '../Common/PostCard/PostCard';
 import PostGalleryItem from './PostGalleryItem';
-import BottomSheetModal from '../Modal/BottomSheetModal';
-import BottomSheetContent from '../Modal/BottomSheetContent';
-import ConfirmModal from '../Modal/ConfirmModal';
-import Button from '../Button/Button';
+import BottomSheetModal from '../../layouts/Modal/BottomSheetModal';
+import BottomSheetContent from '../../layouts/Modal/BottomSheetContent';
+import ConfirmModal from '../../layouts/Modal/ConfirmModal';
+import Button from '../Common/Button/Button';
 import { ReactComponent as PostGalleryIcon } from '../../assets/icons/icon-post-album.svg';
 import { ReactComponent as PostListIcon } from '../../assets/icons/icon-post-list.svg';
 import useModal from '../../hooks/useModal';
