+++ src/layouts/Splash/Splash.jsx
