+++ src/pages/HomePage.jsx
@@ -2,16 +2,16 @@ import React, { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 import authInstance from '../api/instance/authInstance';
-import Button from '../components/Button/Button';
-import Header from '../components/Header/Header';
-import NavBar from '../components/NavBar/NavBar';
-import PostCard from '../components/Post/PostCard';
-import BottomSheetModal from '../components/Modal/BottomSheetModal';
-import BottomSheetContent from '../components/Modal/BottomSheetContent';
-import ConfirmModal from '../components/Modal/ConfirmModal';
+import Button from '../components/Common/Button/Button';
+import Header from '../layouts/Header/Header';
+import NavBar from '../layouts/NavBar/NavBar';
+import PostCard from '../components/Common/PostCard/PostCard';
+import BottomSheetModal from '../layouts/Modal/BottomSheetModal';
+import BottomSheetContent from '../layouts/Modal/BottomSheetContent';
+import ConfirmModal from '../layouts/Modal/ConfirmModal';
 import useModal from '../hooks/useModal';
 import logo from '../assets/images/logo.png';
-import Loading from '../components/Loading/Loading';
+import Loading from '../layouts/Loading/Loading';
 
 const ContentWrapper = styled.main`
   margin: 48px 0 0;
