+++ src/pages/ProductUploadPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Header from '../components/Header/Header';
+import Header from '../layouts/Header/Header';
 import ProductUpload from '../components/Product/ProductUpload';
 import authInstance from '../api/instance/authInstance';
